Add support for uploading and downloading documents.
On the upload side, data is streamed to the server using
a multi-part form request.

On the download side, data is streamed to the client and
the client is provided with the stream so that it can
efficiently process the file (e.g. stream to disk, etc.)

Added both mock and integration unit tests.